extract ZMQ functionality into a proxy object
